+++ src/__tests__/extensions/sessionrecording.js
@@ -164,6 +164,7 @@ describe('SessionRecording', () => {
                 slimDOMOptions: {},
                 collectFonts: false,
                 plugins: [],
+                inlineStylesheet: true,
             })
         })