+++ src/extensions/sessionrecording.js
@@ -130,6 +130,7 @@ export class SessionRecording {
             maskInputFn: null,
             slimDOMOptions: {},
             collectFonts: false,
+            inlineStylesheet: true,
         }
         // We switched from loading all of rrweb to just the record part, but
         // keep backwards compatibility if someone hasn't upgraded PostHog