+++ src/posthog-core.js
@@ -14,7 +14,7 @@ import { cookieStore, localStore } from './storage'
 import { RequestQueue } from './request-queue'
 import { CaptureMetrics } from './capture-metrics'
 import { compressData, decideCompression } from './compression'
-import { encodePostData, xhr } from './send-request'
+import { addParamsToURL, encodePostData, xhr } from './send-request'
 import { RetryQueue } from './retry-queue'
 import { SessionIdManager } from './sessionid'
 import { getPerformanceData } from './apm'
@@ -99,6 +99,7 @@ const defaultConfig = () => ({
         maskInputFn: null,
         slimDOMOptions: {},
         collectFonts: false,
+        inlineStylesheet: true,
     },
     mask_all_element_attributes: false,
     mask_all_text: false,
@@ -412,12 +413,9 @@ PostHogLib.prototype._send_request = function (url, data, options, callback) {
     }
 
     const useSendBeacon = window.navigator.sendBeacon && options.transport.toLowerCase() === 'sendbeacon'
-    var args = options.urlQueryArgs || {}
-    args['ip'] = this.get_config('ip') ? 1 : 0
-    args['_'] = new Date().getTime().toString()
-
-    const argSeparator = url.indexOf('?') > -1 ? '&' : '?'
-    url += argSeparator + _.HTTPBuildQuery(args)
+    url = addParamsToURL(url, options.urlQueryArgs, {
+        ip: this.get_config('ip'),
+    })
 
     if (_.isObject(data) && this.get_config('img')) {
         var img = document.createElement('img')
@@ -1205,7 +1203,7 @@ PostHogLib.prototype.sessionRecordingStarted = function () {
  * returns the current config object for the library.
  */
 PostHogLib.prototype.get_config = function (prop_name) {
-    return this['config'][prop_name]
+    return this.config?.[prop_name]
 }
 
 /**