+++ src/send-request.js
@@ -1,4 +1,26 @@
 import { _, logger } from './utils'
+import Config from './config'
+
+export const addParamsToURL = (url, urlQueryArgs, parameterOptions) => {
+    const args = urlQueryArgs || {}
+    args['ip'] = parameterOptions['ip'] ? 1 : 0
+    args['_'] = new Date().getTime().toString()
+    args['ver'] = Config.LIB_VERSION
+
+    const halves = url.split('?')
+    if (halves.length > 1) {
+        const params = halves[1].split('&')
+        for (const p of params) {
+            const key = p.split('=')[0]
+            if (args[key]) {
+                delete args[key]
+            }
+        }
+    }
+
+    const argSeparator = url.indexOf('?') > -1 ? '&' : '?'
+    return url + argSeparator + _.HTTPBuildQuery(args)
+}
 
 export const encodePostData = (data, options) => {
     if (options.blob && data.buffer) {
@@ -51,6 +73,7 @@ export const xhr = ({
     _.each(headers, function (headerValue, headerName) {
         req.setRequestHeader(headerName, headerValue)
     })
+
     if (options.method === 'POST' && !options.blob) {
         req.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded')
     }